Added cool colouring depending on elevation gain, but tbh doesn't loo…

+++ SimpleCPXWrapper/src/main/cpp/include/Location.h
@@ -11,7 +11,7 @@ typedef struct Location Location;
 struct Location {
   double latitude;
   double longitude;
-  long elevation;
+  double elevation;
   unsigned long long time;
   Location *next;
 };
